feat(view): recieve input for move direction
